Move Kerberos implementation to its own module
Henceforth to use Kerberos authentication, you must add this import
statement somewhere in your program:

import _ "github.com/lib/pq/auth/kerberos"

Moving Kerberos to its own module reduces the number of dependencies
for those users who don't need Kerberos.

Closes: #971
