+++ krb.go
@@ -1,40 +1,24 @@
 package pq
 
-import (
-	"net"
-	"strings"
-)
+// A function that creates a GSS authentication provider. You
+// only need to care about this type if you are writing a GSS
+// authentication provider.
+type NewGSSFunc func() (Gss, error)
 
-/*
- *  Basic GSSAPI interface to abstract Windows (SSPI) from Unix
- *  APIs within the driver
- */
+var newGss NewGSSFunc
 
+// Register the function for creating a GSS authentication provider.
+// You only need to care about this function if you are writing a
+// GSS authentication provider.
+func RegisterNewGSSFunc(newGssArg NewGSSFunc) {
+	newGss = newGssArg
+}
+
+// An interface for providing GSSAPI authentication (e.g. Kerberos).
+// You only need to care about this interface if you are writing a
+// GSS authentication provider.
 type Gss interface {
 	GetInitToken(host string, service string) ([]byte, error)
 	GetInitTokenFromSpn(spn string) ([]byte, error)
 	Continue(inToken []byte) (done bool, outToken []byte, err error)
-}
-
-/*
- *  Find the A record associated with a hostname
- *  In general, hostnames supplied to the driver should be
- *  canonicalized because the KDC usually only has one
- *  principal and not one per potential alias of a host.
- */
-func canonicalizeHostname(host string) (string, error) {
-	canon := host
-
-	name, err := net.LookupCNAME(host)
-	if err != nil {
-		return "", err
-	}
-
-	name = strings.TrimSuffix(name, ".")
-
-	if name != "" {
-		canon = name
-	}
-
-	return canon, nil
 }
